requested changes added

+++ web/src/main/java/com/infoshareacademy/jjdd6/servlet/ShareServlet.java
@@ -17,6 +17,7 @@ import javax.servlet.http.HttpServletResponse;
 import javax.transaction.Transactional;
 import java.io.IOException;
 import java.math.BigDecimal;
+import java.time.DateTimeException;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.List;
@@ -38,12 +39,8 @@ public class ShareServlet extends HttpServlet {
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws IOException {
 
-        final String action = req.getParameter("action");
-        LOG.info("Requested action: {}", action);
-        if (action == null || action.isEmpty()) {
-            resp.getWriter().write("Empty action parameter.");
-            return;
-        }
+        final String action = getString(req, resp);
+        if (action == null) return;
 
         if (action.equals("findAll")) {
             findAll(req, resp);
@@ -58,15 +55,27 @@ public class ShareServlet extends HttpServlet {
         }
     }
 
+    private String getString(HttpServletRequest req, HttpServletResponse resp) throws IOException {
+        final String action = req.getParameter("action");
+        LOG.info("Requested action: {}", action);
+        if (action == null || action.isEmpty()) {
+            resp.getWriter().write("Empty action parameter.");
+        }
+        return action;
+    }
+
     private void updateShare(HttpServletRequest req, HttpServletResponse resp)
             throws IOException {
-        final Long id = Long.parseLong(req.getParameter("id"));
-        LOG.info("Updating share with id = {}", id);
+        try {
+            final Long id = Long.parseLong(req.getParameter("id"));
+            LOG.info("Updating share with id = {}", id);
 
-        final Share existingShare = shareDao.findById(id);
-        if (existingShare == null) {
-            LOG.info("No share found for id = {}, nothing to be updated", id);
-        } else {
+            final Share existingShare = shareDao.findById(id);
+            if (existingShare == null) {
+                LOG.info("No share found for id = {}, nothing to be updated", id);
+                return;
+            }
+
             String ticker = req.getParameter("ticker");
             existingShare.setTicker(ticker);
 
@@ -74,54 +83,100 @@ public class ShareServlet extends HttpServlet {
             existingShare.setFullCompanyName(compName);
 
             String currentPriceStr = req.getParameter("current-price");
-            Long currentPrice = Long.parseLong(currentPriceStr);
-            existingShare.setCurrentPrice(BigDecimal.valueOf(currentPrice));
+            try {
+                Double currentPrice = Double.parseDouble(currentPriceStr);
+                existingShare.setCurrentPrice(BigDecimal.valueOf(currentPrice));
+            } catch (NumberFormatException e) {
+                resp.getWriter().println("Current price should be a number");
+            }
 
             String takeProfitPriceStr = req.getParameter("take-profit");
-            Long takeProfitPrice = Long.parseLong(takeProfitPriceStr);
-            existingShare.setTakeProfitPrice(BigDecimal.valueOf(takeProfitPrice));
+            try {
+                Double takeProfitPrice = Double.parseDouble(takeProfitPriceStr);
+                existingShare.setTakeProfitPrice(BigDecimal.valueOf(takeProfitPrice));
+            } catch (NumberFormatException e) {
+                resp.getWriter().println("Take profit price should have a number value");
+            }
 
             String stopLossPriceStr = req.getParameter("stop-loss");
-            Long stopLossPrice = Long.parseLong(stopLossPriceStr);
-            existingShare.setStopLossPrice(BigDecimal.valueOf(stopLossPrice));
-
+            try {
+                Double stopLossPrice = Double.parseDouble(stopLossPriceStr);
+                existingShare.setStopLossPrice(BigDecimal.valueOf(stopLossPrice));
+            } catch (NumberFormatException e) {
+                resp.getWriter().println("Stop loss price should have a number value");
+            }
             String currentPEstr = req.getParameter("current-PE");
-            Double currentPE = Double.parseDouble(currentPEstr);
-            existingShare.setCurrentPE(currentPE);
-
+            try {
+                Double currentPE = Double.parseDouble(currentPEstr);
+                existingShare.setCurrentPE(currentPE);
+            } catch (NumberFormatException e) {
+                resp.getWriter().println("Current PE should have a number value");
+            }
             String volumeStr = req.getParameter("volume");
-            Long volume = Long.parseLong(volumeStr);
-            existingShare.setVolume(volume);
-
+            try {
+                Long volume = Long.parseLong(volumeStr);
+                existingShare.setVolume(volume);
+            } catch (NumberFormatException e) {
+                resp.getWriter().println("Volume value should be na integer");
+            }
             String highestPriceStr = req.getParameter("highest-price");
-            Long highestPrice = Long.parseLong(highestPriceStr);
-            existingShare.setHighestPrice(BigDecimal.valueOf(highestPrice));
-
+            try {
+                Double highestPrice = Double.parseDouble(highestPriceStr);
+                existingShare.setHighestPrice(BigDecimal.valueOf(highestPrice));
+            } catch (NumberFormatException e) {
+                resp.getWriter().println("Highest price should be a number");
+            }
             String lowestPriceStr = req.getParameter("lowest-price");
-            Long lowestPrice = Long.parseLong(lowestPriceStr);
-            existingShare.setLowestPrice(BigDecimal.valueOf(lowestPrice));
-
+            try {
+                Double lowestPrice = Double.parseDouble(lowestPriceStr);
+                existingShare.setLowestPrice(BigDecimal.valueOf(lowestPrice));
+            } catch (NumberFormatException e) {
+                resp.getWriter().println("Lowest price should be a number");
+            }
             String dateStr = req.getParameter("data-date");
-            LocalDate date = LocalDate.parse(dateStr);
-            existingShare.setDataDate(date);
-
+            try {
+                LocalDate date = LocalDate.parse(dateStr);
+                existingShare.setDataDate(date);
+            } catch (DateTimeException e) {
+                resp.getWriter().println("Date should be: yyyy-mm-dd");
+            }
             String timeStr = req.getParameter("time");
-            LocalTime time = LocalTime.parse(timeStr);
-            existingShare.setDataTime(time);
-
+            try {
+                LocalTime time = LocalTime.parse(timeStr);
+                existingShare.setDataTime(time);
+            } catch (DateTimeException e) {
+                resp.getWriter().println("Time should be hh:mm");
+            }
             String walletIdStr = req.getParameter("wallet-id");
-            Long walletId = Long.parseLong(walletIdStr);
-            Wallet wallet = walletDao.findById(walletId);
-            existingShare.getWallets().add(wallet);
+            try {
+                Long walletId = Long.parseLong(walletIdStr);
+                Wallet wallet = walletDao.findById(walletId);
+                if (wallet != null) {
+                    existingShare.getWallets().add(wallet);
+                } else {
+                    resp.getWriter().println("There is no wallet with such id");
+                }
+            } catch (NumberFormatException e) {
+                resp.getWriter().println("Walletid should be an integer");
+            }
 
             String transactionIdStr = req.getParameter("transaction-id");
-            Long transactionId = Long.parseLong(transactionIdStr);
-            Transaction transaction = transactionDao.findById(transactionId);
-            existingShare.getTransactionLinkedList().add(transaction);
-            existingShare.getTransactionHistory().add(transaction);
-
+            try {
+                Long transactionId = Long.parseLong(transactionIdStr);
+                Transaction transaction = transactionDao.findById(transactionId);
+                if (transaction == null) {
+                    LOG.info("No transaction found for id = {}, nothing to be added to the list", id);
+                } else {
+                    existingShare.getTransactionLinkedList().add(transaction);
+                    existingShare.getTransactionHistory().add(transaction);
+                }
+            } catch (NumberFormatException e) {
+                resp.getWriter().println("Transaction id should be an integer");
+            }
             shareDao.update(existingShare);
             LOG.info("Share object updated: {}", existingShare);
+        } catch (NumberFormatException e) {
+            resp.getWriter().println("Share id should be an integer");
         }
 
         findAll(req, resp);
@@ -141,9 +196,17 @@ public class ShareServlet extends HttpServlet {
 
     private void deleteShare(HttpServletRequest req, HttpServletResponse resp) throws IOException {
 
-        final Long id = Long.parseLong(req.getParameter("id"));
-        LOG.info("Removing Share with id = {}", id);
-        shareDao.delete(id);
+        try {
+            final Long id = Long.parseLong(req.getParameter("id"));
+            if (shareDao.findById(id) != null) {
+                LOG.info("Removing Share with id = {}", id);
+                shareDao.delete(id);
+            } else {
+                resp.getWriter().println("There is no share with id = " + id);
+            }
+        } catch (NumberFormatException e) {
+            resp.getWriter().println("Share id should be an integer");
+        }
 
         findAll(req, resp);
     }
